restructuring
not sure if i'm happy with this but i've messsed around with lifetimes
for long enough to have arrived at a 'good enough' point now.
