+++ src/app.rs
@@ -1,154 +1,43 @@
-use std::borrow::Cow;
+mod graphics;
+
 use winit::{
     event::{Event, WindowEvent},
     event_loop::EventLoop,
-    window::Window,
+    window::WindowBuilder,
 };
 
 pub struct App {}
 
 impl App {
-    pub async fn run(event_loop: EventLoop<()>, window: Window) {
-        let mut size = window.inner_size();
-        size.width = size.width.max(1);
-        size.height = size.height.max(1);
-
-        let instance = wgpu::Instance::default();
-
-        let surface = instance.create_surface(&window).unwrap();
-        let adapter = instance
-            .request_adapter(&wgpu::RequestAdapterOptions {
-                power_preference: wgpu::PowerPreference::default(),
-                force_fallback_adapter: false,
-                compatible_surface: Some(&surface),
-            })
-            .await
-            .expect("Failed to find an appropriate adapter");
-
-        let (device, queue) = adapter
-            .request_device(
-                &wgpu::DeviceDescriptor {
-                    label: None,
-                    required_features: wgpu::Features::empty(),
-                    required_limits: wgpu::Limits::default(),
-                },
-                None,
-            )
-            .await
-            .expect("Failed to create device");
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: None,
-            source: wgpu::ShaderSource::Wgsl(Cow::Borrowed(include_str!(
-                "assets/shaders/shader.wgsl"
-            ))),
-        });
-
-        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: None,
-            bind_group_layouts: &[],
-            push_constant_ranges: &[],
-        });
-
-        // let swapchain_capabilities = surface.get_capabilities(&adapter);
-        // let swapchain_format = swapchain_capabilities.formats[0];
-
-        let mut config = surface
-            .get_default_config(&adapter, size.width, size.height)
-            .unwrap();
-        surface.configure(&device, &config);
-
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: None,
-            layout: Some(&pipeline_layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                // targets: &[Some(swapchain_format.into())],
-                targets: &[Some(wgpu::ColorTargetState {
-                    format: config.format,
-                    blend: Some(wgpu::BlendState::REPLACE),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-            }),
-            primitive: wgpu::PrimitiveState::default(),
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-        });
+    pub async fn run(&mut self) {
+        let event_loop = EventLoop::new().unwrap();
+        let builder = WindowBuilder::new();
+        let window = builder.build(&event_loop).unwrap();
+        let mut app_window = graphics::AppWindow::init(&window).await;
 
         let window = &window;
+        let app_window = &mut app_window;
         event_loop
             .run(move |event, target| {
-                // Have the closure take ownership of the resources.
-                // `event_loop.run` never returns, therefore we must do this to ensure
-                // the resources are properly cleaned up.
-                let _ = (&instance, &adapter);
-
                 if let Event::WindowEvent {
                     window_id: _,
                     event,
                 } = event
                 {
                     match event {
-                        WindowEvent::Resized(new_size) => {
-                            config.width = new_size.width.max(1);
-                            config.height = new_size.height.max(1);
-                            surface.configure(&device, &config);
-
-                            window.request_redraw();
-                        }
+                        WindowEvent::Resized(new_size) => app_window.resize(&window, new_size),
                         WindowEvent::CloseRequested => target.exit(),
-                        WindowEvent::RedrawRequested => {
-                            let frame = surface
-                                .get_current_texture()
-                                .expect("Failed to acquire next swap chain texture");
-                            let view = frame
-                                .texture
-                                .create_view(&wgpu::TextureViewDescriptor::default());
-                            let mut encoder =
-                                device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                                    label: None,
-                                });
-                            {
-                                let mut rpass =
-                                    encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                                        label: None,
-                                        color_attachments: &[Some(
-                                            wgpu::RenderPassColorAttachment {
-                                                view: &view,
-                                                resolve_target: None,
-                                                ops: wgpu::Operations {
-                                                    load: wgpu::LoadOp::Clear(wgpu::Color {
-                                                        r: 0.2,
-                                                        g: 0.28,
-                                                        b: 0.35,
-                                                        a: 1.0,
-                                                    }),
-                                                    store: wgpu::StoreOp::Store,
-                                                },
-                                            },
-                                        )],
-                                        depth_stencil_attachment: None,
-                                        timestamp_writes: None,
-                                        occlusion_query_set: None,
-                                    });
-                                rpass.set_pipeline(&render_pipeline);
-                                rpass.draw(0..3, 0..1);
-                            }
-
-                            queue.submit(Some(encoder.finish()));
-                            frame.present();
-                        }
+                        WindowEvent::RedrawRequested => app_window.render(),
                         _ => {}
                     };
                 }
             })
             .unwrap();
+    }
+
+    pub fn init() -> App {
+        env_logger::init();
+
+        App {}
     }
 }
